backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.5.0
python-dotenv==1.0.1
pydantic==2.7.1
groq==0.9.0
requests==2.32.3
PyGithub==2.3.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
